refactor(controller): create and move some ui components
