+++ src/controller.ts
@@ -50,48 +50,30 @@ import {
 
 import { FOLDER_SORT } from './constants/sort.result.js';
 import ansiEscapes from 'ansi-escapes';
-import { bufferUntil } from './libs/buffer-until.js';
 import { homedir } from 'os';
 import colors from 'colors';
 import keypress from 'keypress';
 import __dirname from './dirname.js';
 import path from 'path';
+import { BannerUi } from './ui/header.ui.js';
+import { UiService } from './services/ui.service.js';
+import { ResultsUi } from './ui/results.ui.js';
+import { GeneralUi } from './ui/general.ui.js';
+import { HelpUi } from './ui/help.ui.js';
 
 export class Controller {
   private folderRoot = '';
-  private stdin: NodeJS.ReadStream = process.stdin;
   private stdout: NodeJS.WriteStream = process.stdout;
   private config: IConfig = DEFAULT_CONFIG;
-
-  private cursorPosY = MARGINS.ROW_RESULTS_START;
-  private previusCursorPosY = MARGINS.ROW_RESULTS_START;
-
-  private scroll: number = 0;
 
   private searchStart: number;
   private searchDuration: number;
   private finishSearching$: Subject<boolean> = new Subject<boolean>();
 
-  private KEYS: IKeysCommand = {
-    up: this.moveCursorUp.bind(this),
-    // tslint:disable-next-line: object-literal-sort-keys
-    down: this.moveCursorDown.bind(this),
-    space: this.delete.bind(this),
-    j: this.moveCursorDown.bind(this),
-    k: this.moveCursorUp.bind(this),
-    h: this.moveCursorPageDown.bind(this),
-    l: this.moveCursorPageUp.bind(this),
-    d: this.moveCursorPageDown.bind(this),
-    u: this.moveCursorPageUp.bind(this),
-    pageup: this.moveCursorPageUp.bind(this),
-    pagedown: this.moveCursorPageDown.bind(this),
-    home: this.moveCursorFirstResult.bind(this),
-    end: this.moveCursorLastResult.bind(this),
+  private uiGeneral: GeneralUi;
+  private uiResults: ResultsUi;
 
-    execute(command: string, params: string[]) {
-      return this[command](params);
-    },
-  };
+  private KEYS: IKeysCommand;
 
   constructor(
     private fileService: FileService,
@@ -99,9 +81,25 @@ export class Controller {
     private consoleService: ConsoleService,
     private updateService: UpdateService,
     private resultsService: ResultsService,
+    private uiService: UiService,
   ) {}
 
   init(): void {
+    const banner = new BannerUi();
+    this.uiResults = new ResultsUi(
+      this.resultsService,
+      this.consoleService,
+      this.fileService,
+    );
+
+    this.uiGeneral = new GeneralUi();
+    if (this.consoleService.isRunningBuild()) {
+      banner.programVersion = this.getVersion();
+    }
+    this.uiService.add(banner);
+    this.uiService.add(this.uiResults);
+    this.uiService.add(this.uiGeneral);
+
     keypress(process.stdin);
     this.setErrorEvents();
     this.getArguments();
@@ -110,6 +108,7 @@ export class Controller {
     this.initializeLoadingStatus();
     if (this.config.checkUpdates) this.checkVersion();
 
+    this.setupKeysCommand();
     this.scan();
   }
 
@@ -129,7 +128,7 @@ export class Controller {
     }
     if (options['sort-by']) {
       if (!this.isValidSortParam(options['sort-by'])) {
-        this.print(INFO_MSGS.NO_VALID_SORT_NAME);
+        new GeneralUi().print(INFO_MSGS.NO_VALID_SORT_NAME);
         process.exit();
       }
       this.config.sortBy = options['sort-by'];
@@ -161,46 +160,39 @@ export class Controller {
     this.folderRoot = this.folderRoot.replace(/[\/\\]$/, '');
   }
 
+  private setupKeysCommand() {
+    this.KEYS = {
+      up: this.uiResults.moveCursorUp.bind(this),
+      // tslint:disable-next-line: object-literal-sort-keys
+      down: this.uiResults.moveCursorDown.bind(this),
+      space: this.delete.bind(this),
+      j: this.uiResults.moveCursorDown.bind(this),
+      k: this.uiResults.moveCursorUp.bind(this),
+      h: this.uiResults.moveCursorPageDown.bind(this),
+      l: this.uiResults.moveCursorPageUp.bind(this),
+      d: this.uiResults.moveCursorPageDown.bind(this),
+      u: this.uiResults.moveCursorPageUp.bind(this),
+      pageup: this.uiResults.moveCursorPageUp.bind(this),
+      pagedown: this.uiResults.moveCursorPageDown.bind(this),
+      home: this.uiResults.moveCursorFirstResult.bind(this),
+      end: this.uiResults.moveCursorLastResult.bind(this),
+
+      execute(command: string, params: string[]) {
+        return this[command](params);
+      },
+    };
+  }
+
   private showHelp(): void {
-    this.clear();
-    this.print(colors.inverse(INFO_MSGS.HELP_TITLE + '\n\n'));
-    this.print(HELP_HEADER + '\n\n');
-
-    let lineCount = 0;
-    OPTIONS.map((option, index) => {
-      this.printAtHelp(
-        option.arg.reduce((text, arg) => text + ', ' + arg),
-        {
-          x: UI_HELP.X_COMMAND_OFFSET,
-          y: index + UI_HELP.Y_OFFSET + lineCount,
-        },
-      );
-      const description = this.consoleService.splitWordsByWidth(
-        option.description,
-        this.stdout.columns - UI_HELP.X_DESCRIPTION_OFFSET,
-      );
-
-      description.map((line) => {
-        this.printAtHelp(line, {
-          x: UI_HELP.X_DESCRIPTION_OFFSET,
-          y: index + UI_HELP.Y_OFFSET + lineCount,
-        });
-        ++lineCount;
-      });
-    });
-
-    this.printAt(HELP_FOOTER + '\n', {
-      x: 0,
-      y: lineCount * 2 + 2,
-    });
+    new HelpUi(this.consoleService).show();
   }
 
   private showProgramVersion(): void {
-    this.print('v' + this.getVersion());
+    new GeneralUi().print('v' + this.getVersion());
   }
 
   private showObsoleteMessage(): void {
-    this.print(INFO_MSGS.DISABLED);
+    new GeneralUi().print(INFO_MSGS.DISABLED);
   }
 
   private setColor(color: string) {
@@ -224,29 +216,21 @@ export class Controller {
     return packageData.version;
   }
 
-  private clear(): void {
-    this.print(ansiEscapes.clearTerminal);
-  }
-
-  private print(text: string): void {
-    process.stdout.write.bind(process.stdout)(text);
-  }
-
   private prepareScreen(): void {
     this.checkScreenRequirements();
-    this.setRawMode();
-    this.clear();
+    this.uiService.setRawMode();
+    this.uiService.prepareUi();
     this.printUI();
-    this.hideCursor();
+    this.uiService.setCursorVisible(false);
   }
 
   private checkScreenRequirements(): void {
     if (this.isTerminalTooSmall()) {
-      this.print(INFO_MSGS.MIN_CLI_CLOMUNS);
+      new GeneralUi().print(INFO_MSGS.MIN_CLI_CLOMUNS);
       process.exit();
     }
     if (!this.stdout.isTTY) {
-      this.print(INFO_MSGS.NO_TTY);
+      new GeneralUi().print(INFO_MSGS.NO_TTY);
       process.exit();
     }
   }
@@ -266,69 +250,15 @@ export class Controller {
 
   private showNewInfoMessage(): void {
     const message = colors.magenta(INFO_MSGS.NEW_UPDATE_FOUND);
-    this.printAt(message, UI_POSITIONS.NEW_UPDATE_FOUND);
+    this.uiService.printAt(message, UI_POSITIONS.NEW_UPDATE_FOUND);
   }
 
   private isTerminalTooSmall(): boolean {
     return this.stdout.columns <= MIN_CLI_COLUMNS_SIZE;
   }
 
-  private setRawMode(set = true): void {
-    this.stdin.setRawMode(set);
-    process.stdin.resume();
-  }
-
   private printUI(): void {
-    ///////////////////////////
-    // banner and tutorial
-    this.printAt(BANNER, UI_POSITIONS.INITIAL);
-    this.printAt(
-      colors.yellow(colors.inverse(HELP_MSGS.BASIC_USAGE)),
-      UI_POSITIONS.TUTORIAL_TIP,
-    );
-
-    if (this.consoleService.isRunningBuild()) {
-      this.printAt(colors.gray(this.getVersion()), UI_POSITIONS.VERSION);
-    }
-
-    ///////////////////////////
-    // Columns headers
-    this.printAt(colors.gray(INFO_MSGS.HEADER_COLUMNS), {
-      x: this.stdout.columns - INFO_MSGS.HEADER_COLUMNS.length - 4,
-      y: UI_POSITIONS.FOLDER_SIZE_HEADER.y,
-    });
-
-    ///////////////////////////
-    // npkill stats
-    this.printAt(
-      colors.gray(INFO_MSGS.TOTAL_SPACE + DEFAULT_SIZE),
-      UI_POSITIONS.TOTAL_SPACE,
-    );
-    this.printAt(
-      colors.gray(INFO_MSGS.SPACE_RELEASED + DEFAULT_SIZE),
-      UI_POSITIONS.SPACE_RELEASED,
-    );
-  }
-
-  private printAt(message: string, position: IPosition): void {
-    this.setCursorAt(position);
-    this.print(message);
-  }
-
-  private setCursorAt({ x, y }: IPosition): void {
-    this.print(ansiEscapes.cursorTo(x, y));
-  }
-
-  private printAtHelp(message: string, position: IPosition): void {
-    this.setCursorAtHelp(position);
-    this.print(message);
-    if (!/-[a-zA-Z]/.test(message.substring(0, 2)) && message !== '') {
-      this.print('\n\n');
-    }
-  }
-
-  private setCursorAtHelp({ x, y }: IPosition): void {
-    this.print(ansiEscapes.cursorTo(x));
+    this.uiService.renderAll();
   }
 
   private initializeLoadingStatus(): void {
@@ -345,190 +275,26 @@ export class Controller {
   }
 
   private updateStatus(text: string): void {
-    this.printAt(text, UI_POSITIONS.STATUS);
+    this.uiService.printAt(text, UI_POSITIONS.STATUS);
   }
 
   private printFoldersSection(): void {
-    if (this.resultsService.noResultsAfterCompleted) {
-      this.printNoResults();
-      return;
-    }
-    const visibleFolders = this.getVisibleScrollFolders();
-    this.clearLine(this.previusCursorPosY);
-
-    visibleFolders.map((folder: IFolder, index: number) => {
-      const folderRow = MARGINS.ROW_RESULTS_START + index;
-      this.printFolderRow(folder, folderRow);
-    });
+    this.uiResults.render();
   }
 
-  private printNoResults(): void {
-    const message = `No ${colors[DEFAULT_CONFIG.warningColor](
-      this.config.targetFolder,
-    )} found!`;
-    this.printAt(message, {
-      x: Math.floor(this.stdout.columns / 2 - message.length / 2),
-      y: MARGINS.ROW_RESULTS_START + 2,
-    });
-  }
-
-  private printFolderRow(folder: IFolder, row: number) {
-    let { path, lastModification, size } = this.getFolderTexts(folder);
-    const isRowSelected = row === this.getRealCursorPosY();
-
-    lastModification = colors.gray(lastModification);
-    if (isRowSelected) {
-      path = colors[this.config.backgroundColor](path);
-      size = colors[this.config.backgroundColor](size);
-      lastModification = colors[this.config.backgroundColor](lastModification);
-
-      this.paintBgRow(row);
-    }
-
-    if (folder.isDangerous)
-      path = colors[DEFAULT_CONFIG.warningColor](path + '⚠️');
-
-    this.printAt(path, {
-      x: MARGINS.FOLDER_COLUMN_START,
-      y: row,
-    });
-
-    this.printAt(lastModification, {
-      x: this.stdout.columns - MARGINS.FOLDER_SIZE_COLUMN - 6,
-      y: row,
-    });
-
-    this.printAt(size, {
-      x: this.stdout.columns - MARGINS.FOLDER_SIZE_COLUMN,
-      y: row,
-    });
-  }
-
-  private getFolderTexts(folder: IFolder): {
-    path: string;
-    size: string;
-    lastModification: string;
-  } {
-    const folderText = this.getFolderPathText(folder);
-    let folderSize = `${folder.size.toFixed(DECIMALS_SIZE)} GB`;
-    let daysSinceLastModification =
-      folder.modificationTime !== null && folder.modificationTime > 0
-        ? Math.floor(
-            (new Date().getTime() / 1000 - folder.modificationTime) / 86400,
-          ) + 'd'
-        : '--';
-
-    if (folder.isDangerous) daysSinceLastModification = 'xxx';
-
-    // Align to right
-    const alignMargin = 4 - daysSinceLastModification.length;
-    daysSinceLastModification =
-      ' '.repeat(alignMargin > 0 ? alignMargin : 0) + daysSinceLastModification;
-
-    if (!this.config.folderSizeInGB) {
-      const size = this.fileService.convertGBToMB(folder.size);
-      folderSize = `${size.toFixed(DECIMALS_SIZE)} MB`;
-    }
-
-    const folderSizeText = folder.size ? folderSize : '--';
-
-    return {
-      path: folderText,
-      size: folderSizeText,
-      lastModification: daysSinceLastModification,
-    };
-  }
-
-  private paintBgRow(row: number) {
-    const startPaint = MARGINS.FOLDER_COLUMN_START;
-    const endPaint = this.stdout.columns - MARGINS.FOLDER_SIZE_COLUMN;
-    let paintSpaces = '';
-
-    for (let i = startPaint; i < endPaint; ++i) {
-      paintSpaces += ' ';
-    }
-
-    this.printAt(colors[this.config.backgroundColor](paintSpaces), {
-      x: startPaint,
-      y: row,
-    });
-  }
-
-  private getFolderPathText(folder: IFolder): string {
-    let cutFrom = OVERFLOW_CUT_FROM;
-    let text = folder.path;
-    const ACTIONS = {
-      // tslint:disable-next-line: object-literal-key-quotes
-      deleted: () => {
-        cutFrom += INFO_MSGS.DELETED_FOLDER.length;
-        text = INFO_MSGS.DELETED_FOLDER + text;
-      },
-      // tslint:disable-next-line: object-literal-key-quotes
-      deleting: () => {
-        cutFrom += INFO_MSGS.DELETING_FOLDER.length;
-        text = INFO_MSGS.DELETING_FOLDER + text;
-      },
-      'error-deleting': () => {
-        cutFrom += INFO_MSGS.ERROR_DELETING_FOLDER.length;
-        text = INFO_MSGS.ERROR_DELETING_FOLDER + text;
-      },
-    };
-
-    if (ACTIONS[folder.status]) ACTIONS[folder.status]();
-
-    text = this.consoleService.shortenText(
-      text,
-      this.stdout.columns - MARGINS.FOLDER_COLUMN_END,
-      cutFrom,
-    );
-
-    // This is necessary for the coloring of the text, since
-    // the shortener takes into ansi-scape codes invisible
-    // characters and can cause an error in the cli.
-    text = this.paintStatusFolderPath(text, folder.status);
-
-    return text;
-  }
-
-  private paintStatusFolderPath(folderString: string, action: string): string {
-    const TRANSFORMATIONS = {
-      // tslint:disable-next-line: object-literal-key-quotes
-      deleted: (text) =>
-        text.replace(
-          INFO_MSGS.DELETED_FOLDER,
-          colors.green(INFO_MSGS.DELETED_FOLDER),
-        ),
-      // tslint:disable-next-line: object-literal-key-quotes
-      deleting: (text) =>
-        text.replace(
-          INFO_MSGS.DELETING_FOLDER,
-          colors.yellow(INFO_MSGS.DELETING_FOLDER),
-        ),
-      'error-deleting': (text) =>
-        text.replace(
-          INFO_MSGS.ERROR_DELETING_FOLDER,
-          colors.red(INFO_MSGS.ERROR_DELETING_FOLDER),
-        ),
-    };
-
-    return TRANSFORMATIONS[action]
-      ? TRANSFORMATIONS[action](folderString)
-      : folderString;
-  }
+  private printNoResults(): void {}
 
   private clearFolderSection(): void {
-    for (let row = MARGINS.ROW_RESULTS_START; row < this.stdout.rows; row++) {
-      this.clearLine(row);
-    }
+    this.uiResults.clear();
   }
 
   private setupEventsListener(): void {
-    this.stdin.on('keypress', (ch, key) => {
+    this.uiService.stdin.on('keypress', (ch, key) => {
       if (key && key['name']) this.keyPress(key);
     });
 
     this.stdout.on('resize', () => {
-      this.clear();
+      this.uiService.clear();
       this.printUI();
       this.printStats();
       this.printFoldersSection();
@@ -553,14 +319,6 @@ export class Controller {
     process.on('unhandledRejection', (reason: {}) => {
       this.printError(reason['stack']);
     });
-  }
-
-  private hideCursor(): void {
-    this.print(ansiEscapes.cursorHide);
-  }
-
-  private showCursor(): void {
-    this.print(ansiEscapes.cursorShow);
   }
 
   private scan(): void {
@@ -691,110 +449,27 @@ export class Controller {
   }
 
   private quit(): void {
-    this.setRawMode(false);
-    this.clear();
+    this.uiService.setRawMode(false);
+    this.uiService.clear();
     this.printExitMessage();
-    this.showCursor();
+    this.uiService.setCursorVisible(true);
     process.exit();
   }
 
   private printExitMessage(): void {
     const { spaceReleased } = this.resultsService.getStats();
-    const exitMessage = `Space released: ${spaceReleased}\n`;
-    this.print(exitMessage);
+    new GeneralUi().printExitMessage({ spaceReleased });
   }
 
   private getCommand(keyName: string): string {
     return VALID_KEYS.find((name) => name === keyName);
   }
 
-  private isCursorInLowerTextLimit(positionY: number): boolean {
-    const foldersAmmount = this.resultsService.results.length;
-    return positionY < foldersAmmount - 1 + MARGINS.ROW_RESULTS_START;
-  }
-
-  private isCursorInUpperTextLimit(positionY: number): boolean {
-    return positionY > MARGINS.ROW_RESULTS_START;
-  }
-
-  private moveCursorUp(): void {
-    if (this.isCursorInUpperTextLimit(this.cursorPosY)) {
-      this.previusCursorPosY = this.getRealCursorPosY();
-      this.cursorPosY--;
-      this.fitScroll();
-    }
-  }
-
-  private moveCursorDown(): void {
-    if (this.isCursorInLowerTextLimit(this.cursorPosY)) {
-      this.previusCursorPosY = this.getRealCursorPosY();
-      this.cursorPosY++;
-      this.fitScroll();
-    }
-  }
-
-  private moveCursorPageUp(): void {
-    this.previusCursorPosY = this.getRealCursorPosY();
-    const resultsInPage = this.stdout.rows - MARGINS.ROW_RESULTS_START;
-    this.cursorPosY -= resultsInPage - 1;
-    if (this.cursorPosY - MARGINS.ROW_RESULTS_START < 0)
-      this.cursorPosY = MARGINS.ROW_RESULTS_START;
-    this.fitScroll();
-  }
-
-  private moveCursorPageDown(): void {
-    this.previusCursorPosY = this.getRealCursorPosY();
-    const resultsInPage = this.stdout.rows - MARGINS.ROW_RESULTS_START;
-    const foldersAmmount = this.resultsService.results.length;
-    this.cursorPosY += resultsInPage - 1;
-    if (this.cursorPosY - MARGINS.ROW_RESULTS_START > foldersAmmount)
-      this.cursorPosY = foldersAmmount + MARGINS.ROW_RESULTS_START - 1;
-    this.fitScroll();
-  }
-
-  private moveCursorFirstResult(): void {
-    this.previusCursorPosY = this.getRealCursorPosY();
-    this.cursorPosY = MARGINS.ROW_RESULTS_START;
-    this.fitScroll();
-  }
-
-  private moveCursorLastResult(): void {
-    this.previusCursorPosY = this.getRealCursorPosY();
-    this.cursorPosY =
-      MARGINS.ROW_RESULTS_START + this.resultsService.results.length - 1;
-    this.fitScroll();
-  }
-
-  private fitScroll(): void {
-    const shouldScrollUp =
-      this.cursorPosY < MARGINS.ROW_RESULTS_START + this.scroll + 1;
-    const shouldScrollDown =
-      this.cursorPosY > this.stdout.rows + this.scroll - 2;
-    let scrollRequired = 0;
-
-    if (shouldScrollUp)
-      scrollRequired =
-        this.cursorPosY - MARGINS.ROW_RESULTS_START - this.scroll - 1;
-    else if (shouldScrollDown) {
-      scrollRequired = this.cursorPosY - this.stdout.rows - this.scroll + 2;
-    }
-
-    if (scrollRequired) this.scrollFolderResults(scrollRequired);
-  }
-
-  private scrollFolderResults(scrollAmount: number): void {
-    const virtualFinalScroll = this.scroll + scrollAmount;
-    this.scroll = this.clamp(
-      virtualFinalScroll,
-      0,
-      this.resultsService.results.length,
-    );
-    this.clearFolderSection();
-  }
-
   private delete(): void {
     const nodeFolder =
-      this.resultsService.results[this.cursorPosY - MARGINS.ROW_RESULTS_START];
+      this.resultsService.results[
+        this.uiResults.cursorPosY - MARGINS.ROW_RESULTS_START
+      ];
     this.clearErrors();
     this.deleteFolder(nodeFolder);
   }
@@ -827,19 +502,26 @@ export class Controller {
   }
 
   private printError(error: string): void {
-    const errorText = this.prepareErrorMsg(error);
+    const errorText = (() => {
+      const margin = MARGINS.FOLDER_COLUMN_START;
+      const width = this.stdout.columns - margin - 3;
+      return this.consoleService.shortenText(error, width, width);
+    })();
 
-    this.printAt(colors.red(errorText), {
-      x: 0,
-      y: this.stdout.rows,
-    });
-  }
+    throw new Error('TODO implement');
 
-  private prepareErrorMsg(errMessage: string): string {
-    const margin = MARGINS.FOLDER_COLUMN_START;
-    const width = this.stdout.columns - margin - 3;
-    return this.consoleService.shortenText(errMessage, width, width);
+    // TODO create ErrorUi component
+    // this.uiService.printAt(colors.red(errorText), {
+    //   x: 0,
+    //   y: this.stdout.rows,
+    // });
   }
+  //
+  // private prepareErrorMsg(errMessage: string): string {
+  //   const margin = MARGINS.FOLDER_COLUMN_START;
+  //   const width = this.stdout.columns - margin - 3;
+  //   return this.consoleService.shortenText(errMessage, width, width);
+  // }
 
   private printStats(): void {
     const { totalSpace, spaceReleased } = this.resultsService.getStats();
@@ -850,35 +532,16 @@ export class Controller {
     totalSpacePosition.x += INFO_MSGS.TOTAL_SPACE.length;
     spaceReleasedPosition.x += INFO_MSGS.SPACE_RELEASED.length;
 
-    this.printAt(totalSpace, totalSpacePosition);
-    this.printAt(spaceReleased, spaceReleasedPosition);
-  }
-
-  private getVisibleScrollFolders(): IFolder[] {
-    return this.resultsService.results.slice(
-      this.scroll,
-      this.stdout.rows - MARGINS.ROW_RESULTS_START + this.scroll,
-    );
-  }
-
-  private getRealCursorPosY(): number {
-    return this.cursorPosY - this.scroll;
+    this.uiService.printAt(totalSpace, totalSpacePosition);
+    this.uiService.printAt(spaceReleased, spaceReleasedPosition);
   }
 
   private clearErrors(): void {
     const lineOfErrors = this.stdout.rows;
-    this.clearLine(lineOfErrors);
-  }
-
-  private clearLine(row: number): void {
-    this.printAt(ansiEscapes.eraseLine, { x: 0, y: row });
+    this.uiService.clearLine(lineOfErrors);
   }
 
   private getUserHomePath(): string {
     return homedir();
-  }
-
-  private clamp(num: number, min: number, max: number) {
-    return Math.min(Math.max(num, min), max);
   }
 }
